// src/models/Users.ts
import { Carts } from "./Carts";
import { Roles } from "./Roles";

export class Users {
    id:number;
    birthday: Date;
    gender: Number;
    avatar:String;
    address: String;
    phone: String;
    email: String;
    password: String;
    firstname: String;
    lastname: String;
    description: String;
    role:number;
    status:number;
}